Add the descriptions of the new folders
As the archive of the NumPy meetings notes grows, it's best to dedicate a separate folder for each type of meetings. 
I've opened the following folders: 
- community_meetings
- docs_team_meetings
- web_team_meetings
and adding a description for each of them to readme.md. 
Additionally, I’m adding a description of the “triage_meetings” folder that was open earlier.